cast to the actual type the parameter is using

+++ InstrBufferx64.cpp
@@ -70,11 +70,11 @@ void InstrBufferx64::push_modrm(uint8_t mod, uint8_t regop, uint8_t rm) {
 }
 
 void InstrBufferx64::push_modrm(uint8_t mod, uint8_t regop, Register rm) {
-    push_modrm(mod, regop, static_cast<int>(rm));
+    push_modrm(mod, regop, static_cast<uint8_t>(rm));
 }
 
 void InstrBufferx64::push_modrm(uint8_t mod, Register regop, Register rm) {
-    push_modrm(mod, static_cast<int>(regop), static_cast<int>(rm));
+    push_modrm(mod, static_cast<uint8_t>(regop), static_cast<uint8_t>(rm));
 }
 
 void InstrBufferx64::push_byte(uint8_t byte) {
